requirements.txt
# Latest versions as of February 2024

# FastAPI and Server
fastapi~=0.109.2
uvicorn[standard]~=0.27.1
uvloop==0.21.0
httptools==0.6.4
gunicorn==23.0.0
websockets==14.0

# Security

pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
cryptography~=42.0.4
oqs==0.10.2

# Data Processing
numpy~=1.26.3
sympy~=1.12
mpmath==1.3.0

# Performance and Caching
redis==5.2.0
cachetools==5.5.0
prometheus-client~=0.19.0
statsd==4.0.1

# Monitoring and Logging
python-json-logger==2.0.7
structlog==24.1.0
elastic-apm==6.19.0
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation-fastapi==0.49b1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
hypothesis==6.96.1
faker==22.6.0

# Development Tools
black==24.3.0
isort==5.13.2
mypy==1.8.0
flake8==7.0.0
pylint==3.0.3

# Configuration and Environment
python-dotenv==1.0.1
PyYAML==6.0.1
toml==0.10.2

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.32.3
aiohttp==3.10.2
blake3==0.4.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.7
pdoc3==0.10.0

# Quantum-related
qiskit==1.0.2
pennylane==0.39.0

# NATS and Messaging
nats-py==2.9.0
asyncio-nats-client==0.11.5

# Vault Integration
hvac==2.1.0

# Optional Optimizations
ujson==5.9.0
orjson==3.9.15
msgpack==1.0.7
DateTime~=5.5
pydantic-settings~=2.6.1
Jinja2~=3.1.4
setuptools~=75.1.0
joblib~=1.4.2
scipy~=1.14.1
python-jose[cryptography]==3.3.0  # JWT handling
falcon==3.1.3                     # Fast API support
